# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat API Gateway",
    "description": "\n    A FastAPI backend that provides a clean REST interface for the LangXchange chat service.\n\n    ## Features\n\n    - **Authentication**: Manage authentication with the external chat API\n    - **Sessions**: Create and manage chat sessions with AI agents\n    - **Messages**: Send messages and receive AI responses\n    - **WebSocket**: Real-time chat via WebSocket\n    - **Configuration**: Switch between external/internal modes at runtime\n    - **Convenience Endpoint**: Quick chat endpoint for simple interactions\n\n    ## Getting Started\n\n    1. Use `/auth/login` to authenticate (or let the API auto-authenticate)\n    2. Select an agent with `POST /agents/select?agent_uuid=...`\n    3. Create a session with `POST /sessions`\n    4. Send messages with `POST /sessions/{session_uuid}/messages`\n    5. Or use `POST /chat` for quick stateless interactions\n    6. Or connect via WebSocket at `/ws/chat/{session_uuid}`\n\n    ## Runtime State\n    After selecting an agent and creating a session, all endpoints will\n    automatically use those values \u2014 no need to pass them repeatedly.\n    Check current state with `GET /config/state`.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/agents": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "List Agents",
        "description": "List all available agents for the current application.\n\nReturns a list of agents that can be used for chat sessions.",
        "operationId": "list_agents_agents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to list agents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agents/select": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Select Agent",
        "description": "Select an agent to use for subsequent requests.\n\nSets the runtime agent UUID \u2014 all endpoints that use agent_uuid\nwill use this value until changed.",
        "operationId": "select_agent_agents_select_post",
        "parameters": [
          {
            "name": "agent_uuid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The agent UUID to select",
              "title": "Agent Uuid"
            },
            "description": "The agent UUID to select"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Agent not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to select agent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/config/mode": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "Get Mode",
        "description": "Get the current chat mode.",
        "operationId": "get_mode_config_mode_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatModeResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Configuration"
        ],
        "summary": "Switch Mode",
        "description": "Switch between external and internal chat modes.",
        "operationId": "switch_mode_config_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/config/state": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "Get Runtime State",
        "description": "Get the current runtime state (active agent, session, mode).",
        "operationId": "get_runtime_state_config_state_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStateResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Configuration"
        ],
        "summary": "Set Runtime State",
        "description": "Manually override runtime agent UUID and/or session UUID.\n\nUse this to pin a specific agent or session without going through\nthe select agent / create session flow.",
        "operationId": "set_runtime_state_config_state_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "agent_uuid": {
            "type": "string",
            "title": "Agent Uuid"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "agent_uuid"
        ],
        "title": "AgentResponse",
        "description": "Response model for agent information."
      },
      "AgentsListResponse": {
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentResponse"
            },
            "type": "array",
            "title": "Agents"
          }
        },
        "type": "object",
        "required": [
          "agents"
        ],
        "title": "AgentsListResponse",
        "description": "Response model for listing agents."
      },
...
      "ChatModeRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "external",
              "internal"
            ],
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "ChatModeRequest"
      },
      "ChatModeResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "external",
              "internal"
            ],
            "title": "Mode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "message"
        ],
        "title": "ChatModeResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Error response model."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuntimeStateRequest": {
        "properties": {
          "agent_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Uuid"
          },
          "session_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Uuid"
          }
        },
        "type": "object",
        "title": "RuntimeStateRequest"
      },
      "RuntimeStateResponse": {
        "properties": {
          "agent_uuid": {
            "type": "string",
            "title": "Agent Uuid"
          },
          "session_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Uuid"
          },
          "chat_mode": {
            "type": "string",
            "enum": [
              "external",
              "internal"
            ],
            "title": "Chat Mode"
          }
        },
        "type": "object",
        "required": [
          "agent_uuid",
          "session_uuid",
          "chat_mode"
        ],
        "title": "RuntimeStateResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}